Move all the compiler related files to a subfolder
