+++ npm/src/browser.ts
@@ -63,7 +63,11 @@ export function getCompilerWorker(workerArg: string | Worker): ICompilerWorker {
 
 export type { ICompilerWorker };
 export { log, type LogLevel };
-export { type Dump, type ShotResult, type VSDiagnostic } from "./common.js";
+export {
+  type Dump,
+  type ShotResult,
+  type VSDiagnostic,
+} from "./compiler/common.js";
 export { type CompilerState } from "./compiler/compiler.js";
 export {
   getAllKatas,
@@ -74,4 +78,4 @@ export {
   type Exercise,
 } from "./katas.js";
 export { default as samples } from "./samples.generated.js";
-export { QscEventTarget } from "./events.js";
+export { QscEventTarget } from "./compiler/events.js";
