+++ npm/src/compiler/common.ts
