--- npm/src/worker-node.ts
@@ -1,1 +1,0 @@
-export * from "./compiler/worker-node.js";
